device_tree: data: Add support for Generating CCI and SMMU configuration for Versal platform
In Versal Platform CCI and SMMU can be configured using isolation editor,
Configuration information is saved in a separate isolation spec source file
(*.iss file) which will be exported to xsa.

Inorder to generate the CCI and SMMU configuration for versal platforms
need to read the *.iss file, Update the tcl's to search for *.iss file in
a given xsa if found read the *.iss file contents and generate the CCI and
SMMU  device-tree syntaxes if mapped in the design.

Added support for below use cases
 1. If coherent flag is set for a peripheral genreate dma-cohernet property.
 2. If virtualized flag is set for a peripheral generate iommus property.
 3. If secure flag is set for a peripheral generate xlnx,secure property.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.kedareswara.rao@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>