common_proc: Update ip2drv_prop logic to use the Config.c_ parameters
During DTS creation for PL DMA's, properties are generated and
additional properties with same name and prefix of "c" are added.
Typically, for all IP's while generating property by using hsi only
CONFIG.C_ is generated and it is considered as user parameters but for
the PL DMA's user parameters with CONFIG.c_ and model parameters with
CONFIG.C_ is generated. Add additional check is added to retrieve user
parameters with prefix CONFIG.c_

Signed-off-by: Abin Joseph <abin.joseph@amd.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>