common: Pass the custom dts file input as a list instead of string
While processing the custom dts files, treat them as a list of entries instead of
string. This helps in passing multiple custom dts files in one go. Moreover,
this is needed to support custom dts file usecase with Windows like paths.
backslashes in the path is treated as a precursor to special character when
treated as a string.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>