common_proc.tcl: Fix interrupt assignment for peripherals connected via xlconcat
The current SDT flow is facing a challenge where multiple peripherals linked
through an xlconcat block are assigned identical interrupt numbers. This issue
stems from the absence of logic to manage the interrupt number updates for the
xlconcat block. To rectify this, the SDT logic has been enhanced to guarantee
that each peripheral connected through the xlconcat block receives a distinct
and appropriate interrupt number, aligning with the connections specified in
the provided design.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>