ddrpsv: Update to handle holes in a given DDR region
The existing logic in ddrpsv tcl has a bug in handling a scenario of multiple
memory segments with in a given DDR region.
For example if the DDR region0 of 2GB(0-0x80000000) has multiple segments with
in it with a hole in between like 0-512M as one segment and 1G-2GB as other
segment, then the current ddrpsv tcl will fail as it ends up in including full
DDR with out considering the hole. So, this patch updates the logic to fix these
types of scenarios.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>