common_proc: Dont generate properties with incorrect syntax
In some cases hsi is providing properties in incorrect format and
generating device tree properties for those are resulting in device tree
compilation failures. So, update SDTGen to handle these scenarios by skipping
the property generation for those in system device tree. Also provide an info
about this to users with a warning.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>