ddrpsv: Correct the logic for global memory node generation
Within the logic of global memory node generation, the memory
block count was getting half twice and is causing issue of only half
the memories listed in global memory node. Fix this by removing
the unneeded division by 2 as it is lareday being taken care in
the following for loop.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>