common_proc: Correct get_ip_type to identify PL ips properly
The get_ip_type proc isnt handling the case where a PL IP is getting
returned with get_mem_ranges but not with get_cells -hier. So, update
the proc to handle this by returning -1 if an IP isnt available with
get_cells -hier so that the caller can treat this as PL IP by default.
Though the IP is available with get_mem_ranges, the report property on
the object returned by get_mem_ranges will not have property of IS_PL
hence return -1 and treat it as PL IP as default behavior.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>