kernel_dtsi: versal2: Add SLCR region to be configured by Xen
When operating under the Xen hypervisor, dom0 must access the MMI_GPU_SLCR
address space to configure the GPU correctly. Add the corresponding node
and explicitly set it to no-map, as it is only required by Xen.
If the system configuration does not include the hypervisor,
this node will have no effect.

Signed-off-by: Parth Gajjar <parth.gajjar@amd.com>
Acked-by: Vishal Sagar <vishal.sagar@amd.com>
Acked-by: Varunkumar Allagadapa <varunkumar.allagadapa@amd.com>