common_proc: Put handle name instead of IP_NAME as the node name for bram_ctrl.
This fix is a workaround. Lately, a design is encountered in Vivado test suite
where there are 2 bram controllers, both starting at same address 0. None of
them are mapped to any processor, their MASTER INTERFACEs are PCI interfaces.
This is leading to duplicate node names (axi_bram_ctrl@0) with different labels
in SDT.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>