device_tree.tcl: Add file checks for sdtgen inputs
SDTGen hangs when XSA directory is given instead of the whole XSA path in set_dt_param. Hence add file checks for -xsa and -rm_xsa to throw error when the whole XSA path is not given. Also add a file check for -board_dts to throw error when files ending with .dtsi are given as input.

Signed-off-by: Dhriti Sree Rompicharla <DhritiSree.Rompicharla@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>