common_proc: Move the comment in switch case to the body of the case
TCl doesnt interpret comments in switch when they are out of case and
doing so may result in error in some cases when the string in the
comment matches with switch variable. So, move the comment into the case.

Signed-off-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>

Acked-for-series: Appana Durga Kedareswara rao <appana.durga.kedareswara.rao@amd.com>