axi_clk_wiz: Update generic compatible string
Refines the logic used to append compatible strings in the device tree
node for clocking wizard IPs. Previously, the script checked for two specific strings:

xlnx,clk-wizard-1.0
xlnx,clkx5-wiz-1.0
If matched, it appended "xlnx,versal-clk-wizard"; otherwise,
it defaulted to "xlnx,clocking-wizard".

This update changes the logic to:

Use a wildcard match (xlnx,clocking-wizard-*)
to cover all future versions of the standard clocking wizard IP.

Append "xlnx,clocking-wizard" if matched, or "xlnx,versal-clk-wizard" as a fallback.

This makes the handling more robust and adaptable to future IP variations,
reducing the need for manual updates as new versions are introduced.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>