device_tree: set generic compatibles for axi_ips
Each time there is an update the compatible property of the ips changes. The driver is not pulled due to this. Hence set generic compatibles for axi-ips to avoid this issue. Also add a proc in common_proc to check and set the dts the node is generated in so avoid node not found issues.

Signed-off-by: Dhriti Sree Rompicharla <DhritiSree.Rompicharla@amd.com>

Acked-for-series: Onkar Harsh <onkar.harsh@amd.com>