common_proc: set common base address format for proper duplicate node identification
IPs having same base address and same ip name are not being identified as duplicate labels since the difference in trailing zeros in the address is picked by the dict exists logic and they are being treated as 2 different nodes. Hence add a check to normalize address formats to avoid this issue.

Signed-off-by: Dhriti Sree Rompicharla <DhritiSree.Rompicharla@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>