device_tree: Check the valid IP Name before processing info for address map
When get_mem_ranges output of a processor is a nested list and is too large,
after certain number of elements, a ... is observed in the HSI output. Add a
join statement to avoid such behavior. Moreover, add a check to validate the
presence of the IP Name before processing the memory object for the address map.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Signed-off-by: Dhriti Sree Rompicharla <DhritiSree.Rompicharla@amd.com>
Acked-by: Appana Durga Kedareswara rao <appana.durga.kedareswara.rao@amd.com>