common: Add support for small variants of Versal 2VE and 2VM SOC family
In some of the small variants of Versal 2VE and 2VM SOC family, the PS11 and MMI
sub-cores of Wizard is combined into a new block PSXT. Add support for the same.

In addition, make the logic to get the number of RPUs in the design generic.
Remove the platform specific hard coding of number of RPU cores.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>