cpu: Add a property "variant" to define the microblaze family
Microblaze family is available under microblaze node in system-device-tree. This
family name becomes more important in the context of new microblaze devices coming
up. Therefore, introduce a new property "variant" in the SDT root node so that
lopper assists don't have to go to the microblaze node to fetch this entry. If
needed this property can be extended to PS devices as well for macro definitions.

While adding above entry, it is observed that the write_dt API doesn't support
adding a property with a list of strings in the root node. Fix the same.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>