device_tree: Improve firmware-name handling for ZynqMP platforms
Improve the handling of the firmware-name property for ZynqMP platforms
in both device_tree.tcl and partial_proc.tcl. The updated logic prioritizes
.bin files over .bit files by first checking for the presence of TYPE == bin
or partial_bin, and falling back to .bit or partial_bit only if a .bin is not
found. Additionally, a file extension check was added before appending .bin
to avoid duplication and ensure proper formatting.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>