dp_rx: Add support for multiple vidphy instances
If there are multiple vidphy (and dp_rx) instances in a design,
the DTG cannot figure out which vidphy to use
to complete the DT properties of the connected dp_rx.

To create a recognizable bond between a vidphy and
the connected dp_rx, they can be placed in the same hierarchical subblock.
This way, their names get common prefixes.

The dp_rx script is updated to recognize such a relationship
and use the vidphy whose name matches best the name of the dp_rx core.

Signed-off-by: Eddy Debaere <eddy.debaere.ext@barco.com>
Signed-off-by: Lakshmi Prasanna Eachuri <lakshmi.prasanna.eachuri@amd.com>

Acked-for-series: Gaddipati, Naveen <naveen.gaddipati@amd.com>