pmc_dt: Add the definition of processor_ip_list to facilitate base address reading for target IPIs
IPI is a very special case. A PMC IPI can surely be connected to other processor
targeted IPIs and those target IPIs are listed as the child-nodes of this
parent PMC IPI. Before commit 83d94db75dd4 ( ipipsu: Fix unit address mismatch
for IPI child nodes), the child-nodes of pmc ipi node were having unit integer
based addresses in the node name (0,1,2 and so on). Therefore the base addresses
of the target IPIs were not calculated. After this change, the exact child node
addresses were expected in the node name which are not mapped to PMC processor
but their base addresses are still needed.

Add the definition of processor_ip_list to facilitate the reading of the base
address for target IPIs.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu