device_tree: Add support for pmc_wwdt for Versal 2VE and 2VM devices
For Versal 2VE and 2VM devices, the IP_NAME for PMC_WWDT is pmc_wwdt
which causes the wdttb tcl not being picked up in the generation of dts.

Add support for pmc_wwdt for Versal 2VE and 2VM devices for the device
tree generation.

Remove timeout-sec property from pmc_wwdt node in Versal 2VE and 2VM
devices as timeout (in milliseconds) is generated from Vivado, only for
PMC_WWDT.

Signed-off-by: Harini T <harini.t@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>