common_proc.tcl: Prepend partial file prefix to interrupt-parent only for RPs
Updates the gen_interrupt_property procedure to prepend the partial
file name to the interrupt controller only if the controller is
associated with a valid reconfigurable partition (RP). This ensures
that for RPs, the interrupt-parent property includes the partial file
prefix, while for static controllers like axi_intc_0, the interrupt-parent
remains unchanged.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
Acked-by: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>