arm64: versal-net: update thermal sensor and node
Update the device tree to integrate the enhanced thermal
monitoring capability for the AI Engine.

A new thermal zone to monitor the AI Engine specifically is
added along with the existing thermal zone, with two trip info.

Also updated the io-channel-cells in sysmon0 node as 1
as per the dt binding.

Signed-off-by: Anish Kadamathikuttiyil Karthikeyan Pillai <anish.kadamathikuttiyil-karthikeyan-pillai@amd.com>

Acked-for-series: Onkar Harsh <onkar.harsh@amd.com>