axi_bram: Make this TCL more modular to cover all the BRAM use cases
The axi bram TCL was not able to support remapping of bram to a different
address for R5. Making it modular helps it support all the possible use cases
including the BRAM with and without ECC and LMB BRAM use cases.

Signed-off-by: Onkar Harsh <onkar.harsh@amd.com>

Acked-for-series: Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@am=