pmc_dt: Add a check to validate pmc mapped IPs
There are telluride design where there are no IPs mapped to the PMC processor. There is no proper error thrown from SDT side to avoid further processing of the SDT for such cases. Hence update the code to throw an error if there are no IPs mapped to PMC processor in a provided design.

Signed-off-by: Dhriti Sree Rompicharla <DhritiSree.Rompicharla@amd.com>
Acked-by: Onkar Harsh <onkar.harsh@amd.com>